修改

+++ src/main/java/com/mawkun/laundry/controller/AdminController.java
@@ -2,9 +2,12 @@ package com.mawkun.laundry.controller;
 
 import cn.pertech.common.abs.BaseController;
 import cn.pertech.common.spring.JsonResult;
+import com.mawkun.laundry.base.data.UserSession;
 import com.mawkun.laundry.base.entity.Admin;
 import com.mawkun.laundry.service.AdminServiceExt;
+import com.mawkun.laundry.spring.annotation.LoginedAuth;
 import io.swagger.annotations.ApiImplicitParam;
+import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -28,38 +31,44 @@ public class AdminController extends BaseController {
     }
 
     @RequestMapping("/get")
-    public JsonResult getByEntity(Admin admin) {
+    public JsonResult getByEntity(@LoginedAuth UserSession session, Admin admin) {
+        if(session.getShopId() > 0) admin.setId(session.getId());
         Admin resultAdmin = adminServiceExt.getByEntity(admin);
         return sendSuccess(resultAdmin);
     }
 
     @RequestMapping("/list")
-    public JsonResult list(Admin admin) {
+    public JsonResult list(@LoginedAuth UserSession session, Admin admin) {
+        if(session.getShopId() > 0) admin.setId(session.getId());
         List<Admin> adminList = adminServiceExt.listByEntity(admin);
         return sendSuccess(adminList);
     }
 
     @RequestMapping("/insert")
-    public JsonResult insert(@RequestBody Admin admin){
+    public JsonResult insert(@LoginedAuth UserSession session, Admin admin){
+        if(session.getShopId() > 0) return sendSuccess("子管理员无权添加管理员，请联系主管理员添加");
         adminServiceExt.insert(admin);
         return sendSuccess(admin);
     }
 
     @RequestMapping("/update")
-    public JsonResult update(@RequestBody Admin admin){
+    public JsonResult update(@LoginedAuth UserSession session, Admin admin){
+        if(session.getShopId() > 0) admin.setId(session.getId());
         adminServiceExt.update(admin);
         return sendSuccess("编辑成功");
     }
 
-    @RequestMapping("/delete/{id}")
-    public JsonResult deleteOne(@PathVariable Long id){
+    @RequestMapping("/delete")
+    public JsonResult deleteOne(@LoginedAuth UserSession session, Long id){
+        if(session.getShopId() > 0) return sendSuccess("子管理员无删除其他管理员权限");
         adminServiceExt.deleteById(id);
         return sendSuccess("删除成功");
     }
 
-    @RequestMapping("/delete")
-    public JsonResult deleteBatch(@RequestBody List<Long> ids){
+    @RequestMapping("/deleteBatch")
+    public JsonResult deleteBatch(@LoginedAuth UserSession session, @RequestBody List<Long> ids){
         int result = 0;
+        if(session.getShopId() > 0) return sendSuccess("子管理员无删除其他管理员权限");
         if (ids!=null&&ids.size()>0) result = adminServiceExt.deleteByIds(ids);
         return sendSuccess(result);
     }
